requirements.txt
pandas
numpy
s3fs
pyarrow
python-dotenv
